Memory optimizations
- Modified MotorController::flipMotor() to check bits instead of using a switch case.
- Angles are now saved as ints. We now take the angle from the accelerometer (float), multiply it by 100 and change it to an int. It uses less memory, but we still have an accuracy of two decimals.

+++ arduino_code/robot_controller/mpu_reader.h
@@ -10,7 +10,7 @@ class MPUReader {
 
   private:
     int _ledPin;
-    float _currentAngle;
+    int _currentAngle;
     MPU9250 _mpu;
 
     void calibrate(void);
@@ -18,7 +18,7 @@ class MPUReader {
   public:
     MPUReader(int ledPin);
     void mpuSetup(int address);
-    float updateAngle(void);
+    int updateAngle(void);
 
 };
